Use a separate Toolbar for workflows, remove Event Bus and pass listeners from the parent to child component
